requirements.txt
beautifulsoup4>=4.12.0
markdownify>=0.12.1
requests>=2.31.0
chromadb>=0.4.24
openai>=1.12.0
typer>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.23.2
